feat(astro): update to v4
Removed Image component in favor of new builtin. Redid a lot of code
related to manual <Picture> creation and replaced it with the builtins.

Updated icons to use new library versions. Svelte now pulls them in
from a third party CDN but this is only on the canal page so no big
deal.